fix: retries included the error response body (#13)
When curl retried the request the returned output also included the body
of the failed responses. This leads to parsing errors, even if the
actual token was part of the response.

By using `--fail` instead curl no longer logs the bodies of the failed
requests. Instead we add a manual error message for a nicer experience.

Co-authored-by: Lukas Metzner <lukas.metzner@hetzner-cloud.de>